refactor FixedImage
